refactor: theme typography
