+++ app/src/main/kotlin/com/majorkik/movieboxcompose/ui/MainActivity.kt
@@ -13,7 +13,6 @@ import androidx.core.view.WindowCompat
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import com.majorkik.core.ui.CoreStyle
 import com.majorkik.core.ui.theme.MovieBoxTheme
-import com.majorkik.core.ui.theme.MBTheme
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MainActivity : ComponentActivity() {
@@ -31,7 +30,7 @@ class MainActivity : ComponentActivity() {
                 viewModel.shouldUseDarkTheme.collectAsState(initial = isSystemInDarkTheme())
             val systemUiController = rememberSystemUiController()
 
-            MBTheme(isDark = shouldUseDarkTheme.value) {
+            MovieBoxTheme(isDark = shouldUseDarkTheme.value) {
 
                 SideEffect {
                     systemUiController.setSystemBarsColor(
