+++ core/ui/src/main/kotlin/com/majorkik/core/ui/theme/Color.kt
@@ -6,17 +6,12 @@ import androidx.compose.ui.graphics.Color
 
 val White = Color(0xFFFFFFFF)
 val Alabaster = Color(0xFFEEEEEE)
-val MineShaft = Color(0xFF282828)
 val Charade = Color(0xFF0A0A0A)
 val BlueCrayola = Color(0xFF3772FF)
-val CornflowerBlue = Color(0xFF5F8EFF)
 val RoseMadder = Color(0xFFDF2935)
 val VistaBlue = Color(0xFF8BCDA3)
-val MineShaftLight = Color(0xFF323232)
-val Gallery = Color(0xFFEDEDED)
 val SilverChalice = Color(0xFFAFAFAF)
 val DoveGray = Color(0xFF616161)
-val RedOrange = Color(0xFFFF3737)
 val CodGray = Color(0xFF1D1D1D)
 val Silver = Color(0xFFBBBBBB)
 val Emperor = Color(0xFF555555)
