+++ ui/nav-home/src/main/kotlin/com/majorkik/ui/nav/home/ui/NavHome.kt
@@ -44,7 +44,6 @@ import com.majorkik.core.ui.components.InfiniteListHandler
 import com.majorkik.core.ui.extension.clickableWithSimpleRipple
 import com.majorkik.core.ui.extension.showToast
 import com.majorkik.core.ui.theme.MovieBoxTheme
-import com.majorkik.core.ui.theme.MBTheme
 import com.majorkik.tmdb.api.model.Genre
 import com.majorkik.ui.nav.home.component.GenresSwitch
 import com.majorkik.ui.nav.home.component.HorizontalMovieCard
@@ -88,7 +87,7 @@ internal fun NavHomeContent(viewModel: NavHomeViewModelViewModel, openMovieDetai
 
         Text(
             stringResource(id = StringResource.nav_home_screen_search_desc),
-            style = MovieBoxTheme.typography.bodyMedium,
+            style = MovieBoxTheme.typography.body.medium,
             color = MovieBoxTheme.colors.text.primary,
             textAlign = TextAlign.Center,
             modifier = Modifier.fillMaxWidth()
@@ -118,7 +117,7 @@ internal fun NavHomeContent(viewModel: NavHomeViewModelViewModel, openMovieDetai
 
                 Text(
                     stringResource(StringResource.common_search),
-                    style = MovieBoxTheme.typography.bodyMedium,
+                    style = MovieBoxTheme.typography.body.medium,
                     color = MovieBoxTheme.colors.text.primary
                 )
             }
@@ -225,7 +224,7 @@ fun GenresBlock(
         ) {
             Text(
                 text = stringResource(id = StringResource.common_genres),
-                style = MovieBoxTheme.typography.h3,
+                style = MovieBoxTheme.typography.header.h3,
                 color = MovieBoxTheme.colors.text.primary
             )
 
@@ -259,7 +258,7 @@ internal fun GenresSwitchBox(isMovieGenresSelected: Boolean, onToggleSwitch: () 
     ) {
         Text(
             text = stringResource(id = StringResource.common_movie),
-            style = MovieBoxTheme.typography.titleMedium,
+            style = MovieBoxTheme.typography.header.titleMedium,
             color = if (isMovieGenresSelected) {
                 MovieBoxTheme.colors.text.primary
             } else {
@@ -271,7 +270,7 @@ internal fun GenresSwitchBox(isMovieGenresSelected: Boolean, onToggleSwitch: () 
 
         Text(
             text = stringResource(id = StringResource.common_tv),
-            style = MovieBoxTheme.typography.titleMedium,
+            style = MovieBoxTheme.typography.header.titleMedium,
             color = if (isMovieGenresSelected.not()) {
                 MovieBoxTheme.colors.text.primary
             } else {
@@ -320,7 +319,7 @@ private fun handleSideEffect(sideEffect: NavHomeViewModelSideEffect, context: Co
 @Preview(showBackground = true)
 @Composable
 private fun Preview() {
-    MBTheme(isDark = true) {
+    MovieBoxTheme(isDark = true) {
         Surface(
             color = MovieBoxTheme.colors.background.base, modifier = Modifier
                 .fillMaxWidth()
